refactor: Model 파일 이름 변경, Info AllowsArbitraryLoads 키값 추가, MovieInfo 타입 구현
