+++ BoxOffice/Model/DailyBoxofficeObject.swift
@@ -1,0 +1,10 @@
+//
+//  DailyBoxofficeObject.swift
+//  BoxOffice
+//
+//  Created by 레옹아범 ,Andrew on 2023/03/20.
+//
+
+struct DailyBoxofficeObject: Decodable {
+    let boxOfficeResult: BoxofficeResultObject
+}
